move files related to properties to a dedicated package
