+++ plugins/formal-verification/formver.core/src/org/jetbrains/kotlin/formver/embeddings/properties/PropertyEmbedding.kt
@@ -1,8 +1,8 @@
 /*
- * Copyright 2010-2023 JetBrains s.r.o. and Kotlin Programming Language contributors.
+ * Copyright 2010-2025 JetBrains s.r.o. and Kotlin Programming Language contributors.
  * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
 
-package org.jetbrains.kotlin.formver.embeddings
+package org.jetbrains.kotlin.formver.embeddings.properties
 
 class PropertyEmbedding(val getter: GetterEmbedding?, val setter: SetterEmbedding?)
