+++ plugins/formal-verification/formver.core/src/org/jetbrains/kotlin/formver/embeddings/properties/SetterEmbedding.kt
@@ -3,11 +3,11 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
 
-package org.jetbrains.kotlin.formver.embeddings
+package org.jetbrains.kotlin.formver.embeddings.properties
 
 import org.jetbrains.kotlin.formver.conversion.StmtConversionContext
 import org.jetbrains.kotlin.formver.embeddings.expression.ExpEmbedding
 
-interface GetterEmbedding {
-    fun getValue(receiver: ExpEmbedding, ctx: StmtConversionContext): ExpEmbedding
+interface SetterEmbedding {
+    fun setValue(receiver: ExpEmbedding, value: ExpEmbedding, ctx: StmtConversionContext): ExpEmbedding
 }
