+++ plugins/formal-verification/formver.core/src/org/jetbrains/kotlin/formver/embeddings/types/ClassPredicateBuilder.kt
@@ -6,7 +6,7 @@
 package org.jetbrains.kotlin.formver.embeddings.types
 
 import org.jetbrains.kotlin.formver.conversion.AccessPolicy
-import org.jetbrains.kotlin.formver.embeddings.UserFieldEmbedding
+import org.jetbrains.kotlin.formver.embeddings.properties.UserFieldEmbedding
 import org.jetbrains.kotlin.formver.embeddings.expression.*
 import org.jetbrains.kotlin.formver.linearization.pureToViper
 import org.jetbrains.kotlin.formver.names.DispatchReceiverName
