+++ plugins/formal-verification/formver.core/src/org/jetbrains/kotlin/formver/embeddings/types/TypeInvariantEmbedding.kt
@@ -5,7 +5,7 @@
 
 package org.jetbrains.kotlin.formver.embeddings.types
 
-import org.jetbrains.kotlin.formver.embeddings.FieldEmbedding
+import org.jetbrains.kotlin.formver.embeddings.properties.FieldEmbedding
 import org.jetbrains.kotlin.formver.embeddings.expression.*
 import org.jetbrains.kotlin.formver.embeddings.expression.OperatorExpEmbeddings.Implies
 import org.jetbrains.kotlin.formver.viper.MangledName
